with postgresql

+++ src/main/java/com/demyanovsky/exceptions/UserNotFoundException.java
@@ -8,7 +8,7 @@ public class UserNotFoundException extends RuntimeException{
 
 
 
-    public UserNotFoundException(int userId){
+    public UserNotFoundException(long userId){
         super("User with id: " +userId + " not found");
 }
 }
